put grade logic in to core module
